Selectively return scicat v3 or v4 class
