+++ sdk/python/scilog/scicat.py
@@ -22,7 +22,17 @@ class SciCatRestAPI(HttpClient):
             return token
 
 
-class SciCat:
+class SciCatLegacy:
+    def __init__(self, *args, **kwargs):
+        self.http_client = SciCatRestAPI(*args, **kwargs)
+
+    @property
+    def proposals(self):
+        url = self.http_client.address + "/proposals"
+        return self.http_client.get_request(url, headers=HEADER_JSON)
+
+
+class SciCatNew:
     max_iterations = 1000
 
     def __init__(self, *args, return_options=None, **kwargs):
@@ -39,7 +49,7 @@ class SciCat:
             if iteration > self.max_iterations:
                 raise RuntimeError("Exceeded maximum iterations in proposals_batch")
             proposals = self.http_client.get_request(
-                f"{url}?filter={quote(dumps(filter))}", headers=HEADER_JSON
+                f"{url}?filters={quote(dumps(filter))}", headers=HEADER_JSON
             )
             if not proposals or len(proposals) == 0:
                 break
@@ -51,6 +61,14 @@ class SciCat:
         lazy = self.return_options.get("lazy", False)
         generator = (proposal for batch in self._proposals_batch() for proposal in batch)
         return generator if lazy else list(generator)
+
+
+class SciCat:
+    def __new__(cls, *args, **kwargs):
+        token_prefix = kwargs.get("options", {}).get("token_prefix")
+        if token_prefix:
+            return SciCatNew(*args, **kwargs)
+        return SciCatLegacy(*args, **kwargs)
 
 
 class SciCatAuthError(AuthError):
